Separate ConfigInputEvent from ConfigInputEventFlow.
